# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dhan Algo Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity
      description: Returns recent orders and trades
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get all orders
      description: Fetch all orders from Dhan
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a new order
      description: Place a new order on Dhan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /orders/{orderId}:
    get:
      operationId: getOrderById
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: modifyOrder
      tags: [orders]
      summary: Modify an existing order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyOrderBody"
      responses:
        "200":
          description: Order modified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
    delete:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
  /market/quote:
    post:
      operationId: getMarketQuote
      tags: [market]
      summary: Get market quote data
      description: Fetch OHLC and quote data for securities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketQuoteRequest"
      responses:
        "200":
          description: Market quote data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketQuoteResponse"
  /market/historical:
    post:
      operationId: getHistoricalData
      tags: [market]
      summary: Get historical data
      description: Fetch historical daily OHLC data for a security
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoricalDataRequest"
      responses:
        "200":
          description: Historical data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoricalDataResponse"
  /market/intraday:
    post:
      operationId: getIntradayData
      tags: [market]
      summary: Get intraday minute data
      description: Fetch intraday minute-level data for a security
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntradayDataRequest"
      responses:
        "200":
          description: Intraday data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntradayDataResponse"
  /market/option-chain:
    post:
      operationId: getOptionChain
      tags: [market]
      summary: Get option chain
      description: Fetch option chain data for an underlying security
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionChainRequest"
      responses:
        "200":
          description: Option chain data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptionChainResponse"
  /market/expiry-list:
    post:
      operationId: getExpiryList
      tags: [market]
      summary: Get expiry list
      description: Fetch expiry dates for an underlying security
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpiryListRequest"
      responses:
        "200":
          description: Expiry list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpiryListResponse"
...
components:
  schemas:
...
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [order, trade, strategy_execution]
        action:
          type: string
        symbol:
          type: string
        quantity:
          type: integer
        price:
          type: number
        status:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: string
      required:
        - id
        - type
        - action
        - symbol
        - quantity
        - price
        - status
        - timestamp
    Order:
      type: object
      properties:
        orderId:
          type: string
        correlationId:
          type: string
        orderStatus:
          type: string
        transactionType:
          type: string
        exchangeSegment:
          type: string
        productType:
          type: string
        orderType:
          type: string
        validity:
          type: string
        tradingSymbol:
          type: string
        securityId:
          type: string
        quantity:
          type: integer
        disclosedQuantity:
          type: integer
        price:
          type: number
        triggerPrice:
          type: number
        afterMarketOrder:
          type: boolean
        boProfitValue:
          type: number
        boStopLossValue:
          type: number
        legName:
          type: string
        createTime:
          type: string
        updateTime:
          type: string
        exchangeTime:
          type: string
        drvExpiryDate:
          type: string
        drvOptionType:
          type: string
        drvStrikePrice:
          type: number
        omsErrorCode:
          type: string
        omsErrorDescription:
          type: string
        filledQty:
          type: integer
        averagePrice:
          type: number
      required:
        - orderId
        - orderStatus
        - transactionType
        - exchangeSegment
        - productType
        - orderType
        - securityId
        - quantity
        - price
    PlaceOrderBody:
      type: object
      properties:
        securityId:
          type: string
        exchangeSegment:
          type: string
        transactionType:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: integer
        orderType:
          type: string
          enum: [MARKET, LIMIT, SL, SLM]
        productType:
          type: string
          enum: [INTRA, CNC, MARGIN, CO, BO]
        price:
          type: number
        triggerPrice:
          type: number
        disclosedQuantity:
          type: integer
        afterMarketOrder:
          type: boolean
        validity:
          type: string
          enum: [DAY, IOC]
        boProfitValue:
          type: number
        boStopLossValue:
          type: number
        tag:
          type: string
      required:
        - securityId
        - exchangeSegment
        - transactionType
        - quantity
        - orderType
        - productType
        - price
    ModifyOrderBody:
      type: object
      properties:
        orderType:
          type: string
        quantity:
          type: integer
        price:
          type: number
        triggerPrice:
          type: number
        disclosedQuantity:
          type: integer
        validity:
          type: string
        legName:
          type: string
    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
        status:
          type: string
        message:
          type: string
      required:
        - orderId
        - status
...
    MarketQuoteRequest:
      type: object
      properties:
        securities:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        quoteType:
          type: string
          enum: [ltp, ohlc, full]
      required:
        - securities
        - quoteType
    MarketQuoteResponse:
      type: object
      properties:
        data:
          type: object
      required:
        - data
    HistoricalDataRequest:
      type: object
      properties:
        securityId:
          type: string
        exchangeSegment:
          type: string
        instrumentType:
          type: string
        expiryCode:
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
      required:
        - securityId
        - exchangeSegment
        - instrumentType
        - fromDate
        - toDate
    HistoricalDataResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CandleData"
      required:
        - data
    CandleData:
      type: object
      properties:
        timestamp:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: integer
      required:
        - timestamp
        - open
        - high
        - low
        - close
        - volume
    IntradayDataRequest:
      type: object
      properties:
        securityId:
          type: string
        exchangeSegment:
          type: string
        instrumentType:
          type: string
      required:
        - securityId
        - exchangeSegment
        - instrumentType
    IntradayDataResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CandleData"
      required:
        - data
    OptionChainRequest:
      type: object
      properties:
        underSecurityId:
          type: string
        underExchangeSegment:
          type: string
        expiry:
          type: string
          format: date
      required:
        - underSecurityId
        - underExchangeSegment
        - expiry
    OptionChainResponse:
      type: object
      properties:
        data:
          type: object
      required:
        - data
    ExpiryListRequest:
      type: object
      properties:
        underSecurityId:
          type: string
        underExchangeSegment:
          type: string
      required:
        - underSecurityId
        - underExchangeSegment
    ExpiryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
      required:
        - data
...