# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dhan Algo Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  action: string;
  symbol: string;
  quantity: number;
  price: number;
  status: string;
  timestamp: Date;
  details?: string;
}