# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dhan Algo Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  orderId: string;
  correlationId?: string;
  orderStatus: string;
  transactionType: string;
  exchangeSegment: string;
  productType: string;
  orderType: string;
  validity?: string;
  tradingSymbol?: string;
  securityId: string;
  quantity: number;
  disclosedQuantity?: number;
  price: number;
  triggerPrice?: number;
  afterMarketOrder?: boolean;
  boProfitValue?: number;
  boStopLossValue?: number;
  legName?: string;
  createTime?: string;
  updateTime?: string;
  exchangeTime?: string;
  drvExpiryDate?: string;
  drvOptionType?: string;
  drvStrikePrice?: number;
  omsErrorCode?: string;
  omsErrorDescription?: string;
  filledQty?: number;
  averagePrice?: number;
}